preprocessing_service/app/requirements.txt
fastapi
uvicorn
httpx
qdrant-client
python-multipart
mistralai
python-dotenv